getTotTrades added
I've noticed a lot of my get____(username) methods are pretty much the exact same.  We should be able to consolidate a lot of this in the DataController, much like I did with the mapUsernames method by passing in the named query from the RestController.  I'll look into this in a bit.